fix: break circular import causing KNOWN_CURL_DIRS TDZ crash (#1070)
Move `InstallationMethod` type and `parseInstallationMethod()` from
`upgrade.ts` to `binary.ts` to break the circular dependency:

  upgrade.ts → db/install-info.ts → upgrade.ts (via import type)

Under Bun's module evaluation order, this cycle caused `upgrade.ts` to
be evaluated before `binary.ts` finished initializing, making
`KNOWN_CURL_DIRS` (a top-level const) hit the temporal dead zone.

Changes:
- Move `InstallationMethod` type + `parseInstallationMethod` to
  `binary.ts` (the installation-constants module)
- Re-export from `upgrade.ts` for backward compatibility
- Update `install-info.ts` and `setup.ts` to import from `binary.ts`
- Defer `KNOWN_CURL_PATHS` computation to a lazy getter (defensive
  against future circular imports)
- Fix E2E test fixture to use `tsx` instead of `bun run` for source
  mode (a second TDZ in the db/ layer still affects Bun)

Fixes #1070
